+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.30.1] - 2024-10-09
+### Fixed
+- Fix bug in sessions controller, attempting >= comparison with enumerated types.
+
 ## [2.30.0] - 2024-10-07
 ### Added
 #### BOS options